requirements.txt
interval==1.0.0
librosa==0.10.1
scipy==1.10.1
smplx==0.1.28
transformers==4.33.3
numpy~=1.24.3
matplotlib~=3.5.3
textgrid~=1.5
scikit-learn~=1.3.2
pyrender~=0.1.45
trimesh~=4.0.4
tqdm~=4.66.1
pyyaml~=6.0.1